pin index is positive only:proteus_f4

+++ firmware/controllers/can/can_bench_test.cpp
@@ -89,7 +89,7 @@ static void qcSetEtbState(uint8_t dcIndex, uint8_t direction) {
 }
 
 static void setPin(const CANRxFrame& frame, int value) {
-		int outputIndex = frame.data8[2];
+		size_t outputIndex = frame.data8[2];
 		if (outputIndex >= getBoardMetaOutputsCount()) {
 		  criticalError("QC pin index %d out of range", outputIndex);
 			return;
